feat: migrate UI layer and use cases to modular architecture
- Created core:ui module with reusable components and theme
- Moved use cases from app to core:domain
- Updated imports and dependencies across modules
