--- app/src/main/java/com/example/petbreeds/domain/usecase/GetPetsUseCase.kt
@@ -1,16 +1,0 @@
-package com.example.petbreeds.domain.usecase
-
-import com.example.common.NetworkResult
-import com.example.model.Pet
-import com.example.model.PetType
-import com.example.domain.repository.PetRepository
-import kotlinx.coroutines.flow.Flow
-import javax.inject.Inject
-
-class GetPetsUseCase @Inject constructor(
-    private val repository: PetRepository
-) {
-    operator fun invoke(petType: PetType): Flow<NetworkResult<List<Pet>>> {
-        return repository.getPets(petType)
-    }
-}
