--- app/src/main/java/com/example/petbreeds/domain/usecase/ToggleFavoriteUseCase.kt
@@ -1,12 +1,0 @@
-package com.example.petbreeds.domain.usecase
-
-import com.example.domain.repository.PetRepository
-import javax.inject.Inject
-
-class ToggleFavoriteUseCase @Inject constructor(
-    private val repository: PetRepository
-) {
-    suspend operator fun invoke(petId: String) {
-        repository.toggleFavorite(petId)
-    }
-}
