+++ app/src/main/java/com/example/petbreeds/presentation/favorites/FavoritesScreen.kt
@@ -25,10 +25,10 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.example.model.PetType
-import com.example.petbreeds.domain.usecase.FavoritePetsState
+import com.example.domain.usecase.FavoritePetsState
 import com.example.petbreeds.presentation.components.DrawerContent
-import com.example.petbreeds.presentation.components.TopBar
-import com.example.petbreeds.presentation.components.PetCard
+import com.example.ui.components.PetCard
+import com.example.ui.components.TopBar
 import kotlinx.coroutines.launch
 
 @SuppressLint("DefaultLocale")
