+++ core/domain/src/main/java/com/example/domain/usecase/ToggleFavoriteUseCase.kt
@@ -1,0 +1,12 @@
+package com.example.domain.usecase
+
+import com.example.domain.repository.PetRepository
+import javax.inject.Inject
+
+class ToggleFavoriteUseCase @Inject constructor(
+    private val repository: PetRepository
+) {
+    suspend operator fun invoke(petId: String) {
+        repository.toggleFavorite(petId)
+    }
+}
