+++ core/ui/src/main/java/com/example/ui/theme/Color.kt
@@ -1,0 +1,33 @@
+package com.example.ui.theme
+
+import androidx.compose.ui.graphics.Color
+
+// Modern Cat Theme
+val CatPrimary = Color(0xFF4A4AFF)        // Vibrant blue-purple
+val CatPrimaryDark = Color(0xFF3838CC)    // Darker blue-purple
+val CatSecondary = Color(0xFFFF6B9D)      // Pink accent
+val CatBackground = Color(0xFFF8F9FF)     // Very light blue-white
+val CatSurface = Color(0xFFFFFFFF)        // Pure white
+val CatCard = Color(0xFFF0F2FF)           // Light purple-blue tint
+
+// Modern Dog Theme
+val DogPrimary = Color(0xBA00BFA6)        // Vibrant teal
+val DogPrimaryDark = Color(0xFF00A693)    // Darker teal
+val DogSecondary = Color(0xFF4DD0E1)      // Light cyan
+val DogBackground = Color(0xFFF0FFFF)     // Very light cyan
+val DogSurface = Color(0xFFFFFFFF)        // Pure white
+val DogCard = Color(0xFFE0F7F7)           // Light teal tint
+
+// Common Colors
+val White = Color(0xFFFFFFFF)
+val Black = Color(0xFF1A1A1A)             // Softer black
+val DarkGray = Color(0xFF2E2E2E)          // Dark gray for text
+val LightGray = Color(0xFFF5F5F5)         // Light backgrounds
+val MediumGray = Color(0xFF9E9E9E)        // Secondary text
+val ErrorRed = Color(0xFFFF5252)          // Modern error red
+val Favorite = Color(0xFFFF4081)          // Pink favorite color
+val Success = Color(0xFF4CAF50)           // Success green
+
+// Gradient colors
+val GradientStart = Color(0xFF667eea)
+val GradientEnd = Color(0xFF764ba2)
