+++ settings.gradle.kts
@@ -27,3 +27,4 @@ include(":core:network")
 include(":core:database")
 include(":core:data")
 include(":core:domain")
+include(":core:ui")
